requirements.txt
python-telegram-bot==13.15
requests~=2.27.1
dadata~=24.10.0
python-dotenv==1.1.0